requirements.txt
Pillow: https://github.com/python-pillow/Pillow
ephem
picamera
python-daemon
twilio
unittest
adafruit_ina219
